Visuals, GameStates & Areas
Forth day, I decided to change the perspective of the game to be sure that is more aligned with a top-down view. This gave me some problems with how I was planning the game, so I may need to make some major changes after this. To help me envision the new game, I decided to add art to work with.
New Features:
- Pause state working with current features
- New area transition
- New game art, credit to Kenney.nl

Updates / Changes:
- New camera perspective
- Timers can be paused now
- New game manager scene loading

Next step:
- Work on making the art changes
- Decide on how the new perspective will affect the current and future gameplay